backend/requirements.txt
asgiref==3.8.1
django==5.0.2
sqlparse==0.5.3
django-environ==0.11.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg==3.1.18
psycopg-binary==3.2.6
gunicorn==21.2.0
djangorestframework-simplejwt==5.3.1
pytz==2023.3